+++ src/job_hunter_ai/domain/ports/__init__.py
@@ -1,6 +1,17 @@
 from job_hunter_ai.domain.ports.applier_registry import ApplierRegistry
+from job_hunter_ai.domain.ports.extractor_registry import ExtractorRegistry
 from job_hunter_ai.domain.ports.job_applier import JobApplier
 from job_hunter_ai.domain.ports.job_repository import JobRepository
 from job_hunter_ai.domain.ports.job_source import JobSource
+from job_hunter_ai.domain.ports.requirements_extractor import RequirementsExtractor
+from job_hunter_ai.domain.ports.resume_reader import ResumeReader
 
-__all__ = ["ApplierRegistry", "JobApplier", "JobRepository", "JobSource"]
+__all__ = [
+    "ApplierRegistry",
+    "ExtractorRegistry",
+    "JobApplier",
+    "JobRepository",
+    "JobSource",
+    "RequirementsExtractor",
+    "ResumeReader",
+]